requirements.txt
asgiref==3.11.0
brotli==1.2.0
cffi==2.0.0
charset-normalizer==3.4.4
cssselect2==0.8.0
Django==4.2
django-unfold==0.38.0
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
fonttools==4.61.0
gunicorn==21.2.0
html5lib==1.1
mysqlclient==2.2.0
packaging==25.0
pillow==12.0.0
pycparser==2.23
pydyf==0.11.0
pyphen==0.17.2
pytz==2025.2
reportlab==4.4.5
sqlparse==0.5.4
tinycss2==1.5.1
weasyprint==62.3
webencodings==0.5.1
zopfli==0.4.0
